refactor: Rename the old welcome page to `act page`
Welcome page never made sense for this

Now there is a proper (WIP) welcome page
