+++ ui/welcome_page.py
@@ -1,35 +1,43 @@
-from PyQt6.QtWidgets import QWidget, QVBoxLayout, QGroupBox, QLabel, QPushButton, QHBoxLayout
+from PyQt6.QtWidgets import QWidget, QVBoxLayout, QLabel, QPushButton, QHBoxLayout
 from PyQt6.QtCore import Qt
-from init_instances import inst
 from ui.header import Header
+from init_instances import inst
 
-class WelcomePage(QWidget,):
+# Welcome page
+class WelcomePage(QWidget):
     def __init__(self):
         super().__init__()
-        self.header = Header.header(self)
-        # Welcome page
         self.welcomePage = QWidget()
-        ## Layout and gorup
-        welcomerLayout = QVBoxLayout()
-        welcomerGroup = QGroupBox("")
-        welcomerGroupLayout = QVBoxLayout()
-        buttonsLayout = QHBoxLayout()
-        welcomerGroup.setLayout(welcomerGroupLayout)
-        self.welcomePage.setLayout(welcomerLayout) # Set the welcomepage layout
-        ## Widgets
-        self.welcomerLabel = QLabel()
-        self.welcomerLabel.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        self.installButton = QPushButton('Install')
-        self.updateButton = QPushButton('Check for Updates')
-        self.uninstallButton = QPushButton('Uninstall')
-        self.backButton = QPushButton("Back")
-        self.backButton.setFixedSize(80, 30)
-        ## Add widgets / layouts
-        welcomerLayout.addLayout(self.header)
-        welcomerLayout.addWidget(welcomerGroup)
-        welcomerGroupLayout.addWidget(self.welcomerLabel)
-        buttonsLayout.addWidget(self.installButton)
-        buttonsLayout.addWidget(self.updateButton)
-        buttonsLayout.addWidget(self.uninstallButton)
-        welcomerLayout.addLayout(buttonsLayout)
-        welcomerLayout.insertWidget(0, self.backButton)
+
+        # Layout and groups
+        welcomeLayout = QVBoxLayout()
+        self.welcomePage.setLayout(welcomeLayout)
+
+        # Add header
+        #    self.Header = Header().header()  # Initialize the header
+        #    welcomeLayout.addLayout(self.Header)
+
+        # Welcome label
+        welcomeLabel = QLabel("<b><font size='12'>Welcome to Troppical Installer!</font></b>")
+        welcomeLabel.setAlignment(Qt.AlignmentFlag.AlignCenter)  # Center the text horizontally
+        welcomeLayout.addWidget(welcomeLabel)
+
+        # Create a horizontal layout for the buttons
+        buttonLayout = QHBoxLayout()
+
+        # Manage button
+        self.manageButton = QPushButton("Manage")
+        self.manageButton.setFixedSize(200, 100)  # Set size for large rectangular shape
+
+        # Configure button
+        self.configureButton = QPushButton("Configure")
+        self.configureButton.setFixedSize(200, 100)  # Set size for large rectangular shape
+        self.configureButton.setToolTip("WIP: Coming soon!")
+        self.configureButton.setEnabled(False)  # Set the button to be grayed out
+
+        # Add buttons to the layout
+        buttonLayout.addWidget(self.manageButton)
+        buttonLayout.addWidget(self.configureButton)
+
+        # Add button layout to the main layout
+        welcomeLayout.addLayout(buttonLayout)
